Renamed m7_o3 to m7_o2, added new mission to replace m7_o3
