+++ levels/cloudinary_video/level.json
@@ -23,6 +23,6 @@
     "m4_o1_adding_text","m4_o2_adding_image",
     "m5_o1_borders","m5_o2_video_layer","m5_o3_combining_clips","m5_o4_transitions",
     "m6_o1_manipulating_audio", "m6_o2_layering_audio",
-    "m7_o1_video_player", "m7_o2_video_placeholders", "m7_o3_video_summarization"
+    "m7_o1_video_player", "m7_o2_video_player_transformations", "m7_o3_video_summarization"
   ]
 }
