--- levels/cloudinary_video/objectives/m7_o2_video_placeholders/description.md
@@ -1,21 +1,0 @@
-# Video Placeholders
-
-<div class="aside">
-<h3>AGENDA</h3>
-<ul>
-  <li>...</li>
-</ul>
-</div>
-
-
-Thanks for your help! Blah blah blah 
-
-> <b>TIP:</b> Experiment with URLs without leaving the game, using the <button onclick='window.CloudinaryBrowser.showUrlExplorer();'>Cloudinary Explorer</button>
-
-********************
-
-## Preparation
-blah blah blah 
-4. If the code looks correct, then click _HACK_ to proceed.
-
-
