--- levels/cloudinary_video/objectives/m7_o2_video_placeholders/objective.json
@@ -1,15 +1,0 @@
-{
-  "title": "Video Placeholders",
-  "description": "Description",
-  "show_ide": false,
-  "validation_fields": [
-    {
-      "name": "label1",
-      "type": "prompt",
-      "label": "Press __HACK__!"
-    }
-  ],
-  "rewards": {
-    "xp": 500
-  }
-}
