--- levels/cloudinary_video/objectives/m7_o2_video_placeholders/walkthrough.md
@@ -1,17 +1,0 @@
-# Help Guide
-
-## Objective-specific Resources
-- [Video Player](https://cloudinary.com/documentation/cloudinary_video_player?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021)
-
-## Detailed Instructions
-Details
-
-### Additional Documentation
-
-[Cloudinary Documentation](https://cloudinary.com/documentation?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021)
-
-[Cloudinary Academy](https://training.cloudinary.com?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021)
-
-[Cloudinary CookBook Recipes - Transformation Examples](https://cloudinary.com/cookbook?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021)
-
-[Discord Community](https://discord.gg/CCsubwFbvd)
