+++ levels/cloudinary_video/objectives/m7_o2_video_player_transformations/objective.json
@@ -1,0 +1,20 @@
+{
+  "title": "Video Player Transformations",
+  "description": "Add information to and transform the video using Cloudinary's API and video player!",
+  "show_ide": true,
+  "validation_fields": [
+    {
+      "name": "label1",
+      "type": "prompt",
+      "label": "Press the _▶_ button to start / restart the web server with your code. Then, enter the server URL below and press __HACK__!"
+    },
+    {
+      "name": "serverUrl",
+      "type": "text",
+      "label": "Server URL"
+    }
+  ],
+  "rewards": {
+    "xp": 500
+  }
+}
