+++ levels/cloudinary_video/objectives/m7_o3_video_summarization/description.md
@@ -1,33 +1,21 @@
-# Video Summarization
+# Video Placeholders
 
 <div class="aside">
 <h3>AGENDA</h3>
 <ul>
-  <li>Add a title, subtitle, and description to the included video player.</li>
-  <li>Transform the video by editing the video tag data.</li>
+  <li>...</li>
 </ul>
 </div>
 
 
-Nice work so far! Next, we need to use another **video player** to play our **TwilioQuest Flower video** but with some additional information and transformations. You remember how to do transformations, right? Let's reapply that knowledge in this objective!
+Thanks for your help! Blah blah blah 
+
+> <b>TIP:</b> Experiment with URLs without leaving the game, using the <button onclick='window.CloudinaryBrowser.showUrlExplorer();'>Cloudinary Explorer</button>
 
 ********************
 
 ## Preparation
-Learn more about our Cloudinary video player [here](https://cloudinary.com/documentation/cloudinary_video_player?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021). If you need additional resources, check out the __Help__ tab above!
-
-
-
-
-1. Open the **code editor** by clicking on the _\</> Show Code Editor_ button on the right.
-2. You'll notice a web application's source code in the code box. Fill in the code block that says `// ----- Edit the video properties below ----- //` . Use the example in the link above to help you!
-3. Be sure to add the following attributes:
-  - The sample code's defined variable names should be left as is.
-  - The video should be the TwilioQuest Flower video you've used in previous objectives.
-  - The video player should have a title, subtitle, and description.
-  - The crop should be set to fill, and the video size should be 300x300 pixels.
-4. Set your cloud name under the `// ----- Edit the cloud name below ----- //`
-5. Run the server by pressing the _▶_ button.
-6. If the code looks correct, then click _HACK_ to proceed.
+blah blah blah 
+4. If the code looks correct, then click _HACK_ to proceed.
