+++ levels/cloudinary_video/objectives/m7_o3_video_summarization/example.js
@@ -13,31 +13,15 @@ const server = http.createServer((request, response) => {
   <script src="https://unpkg.com/cloudinary-core@latest/cloudinary-core-shrinkwrap.min.js" type="text/javascript"></script>
   <script src="https://unpkg.com/cloudinary-video-player@1.5.9/dist/cld-video-player.min.js" 
       type="text/javascript"></script>
-  <div>
-    <!--- Edit the video properties below --->
-    <video id="tq-player" 
-    controls
-    class="cld-video-player" 
-    data-cld-source='{ "publicId": "[video-public-id]",
-    "info": {
-      "title": "[title]",
-      "subtitle": "[subtitle]",
-      "description": "[description]"
-      }
-    }'
-    data-cld-transformation='{
-      "c___": "____",
-      "w____": ###,
-      "h_____": ###
-    }'
-    ></video>
-    <!--- Edit the video properties above --->
+  <div style="max-width: px">
+    <video id="doc-player"  controls  muted  class="cld-video-player"></video>
   </div>
   <script id="video-code">
-    <!--- Edit the cloudname below --->
-    var cld = cloudinary.Cloudinary.new({ cloud_name: '[__________]' });
-    var player = cld.videoPlayer("tq-player", {});
-    <!--- Edit the cloudname above --->
+    // ----- Put your JS code below ----- //
+    var cld = cloudinary.[insert code here];
+    var demoplayer = cld.[code];
+    demoplayer.[code];    
+    // ----- Put your JS code above ----- //
   </script>
   </html>
   `);
@@ -49,36 +33,3 @@ server.listen(PORT, () => {
   console.log(`Paste this URL into the hack UI to complete the challenge!`);
   console.log(`You can open this URL in your web browser too to test it out.`);
 });
-
-/*<html>
-  <link href="https://unpkg.com/cloudinary-video-player@1.5.9/dist/cld-video-player.min.css" rel="stylesheet">
-  <script src="https://unpkg.com/cloudinary-core@latest/cloudinary-core-shrinkwrap.min.js" type="text/javascript"></script>
-  <script src="https://unpkg.com/cloudinary-video-player@1.5.9/dist/cld-video-player.min.js" 
-      type="text/javascript"></script>
-  <div>
-    <!--- Edit the video properties below --->
-    <video id="tq-player" 
-    controls
-    class="cld-video-player" 
-    data-cld-source='{ "publicId": "[video-public-id]",
-    "info": {
-      "title": "[title]",
-      "subtitle": "[subtitle]",
-      "description": "[description]"
-      }
-    }'
-    data-cld-transformation='{
-      "c___": "____",
-      "w____": ###,
-      "h_____": ###
-    }'
-    ></video>
-    <!--- Edit the video properties above --->
-  </div>
-  <script id="video-code">
-    <!--- Edit the cloudname below --->
-    var cld = cloudinary.Cloudinary.new({ cloud_name: '[__________]' });
-    var player = cld.videoPlayer("tq-player", {});
-    <!--- Edit the cloudname above --->
-  </script>
-  </html>*/
