+++ levels/cloudinary_video/objectives/m7_o3_video_summarization/objective.json
@@ -1,17 +1,12 @@
 {
-  "title": "Video Info and Transformations",
-  "description": "Add information to and transform the video using Cloudinary's API and video player!",
-  "show_ide": true,
+  "title": "Video Placeholders",
+  "description": "Description",
+  "show_ide": false,
   "validation_fields": [
     {
       "name": "label1",
       "type": "prompt",
-      "label": "Press the _▶_ button to start / restart the web server with your code. Then, enter the server URL below and press __HACK__!"
-    },
-    {
-      "name": "serverUrl",
-      "type": "text",
-      "label": "Server URL"
+      "label": "Press __HACK__!"
     }
   ],
   "rewards": {
