+++ levels/cloudinary_video/objectives/m7_o3_video_summarization/walkthrough.md
@@ -2,15 +2,9 @@
 
 ## Objective-specific Resources
 - [Video Player](https://cloudinary.com/documentation/cloudinary_video_player?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021)
-- [Video Player Titles, Subtitles, and Descriptions](https://cloudinary.com/documentation/video_player_customization?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021#video_titles_subtitles_and_descriptions)
-- [Video Player Transformations (Common Configuration Examples)](https://cloudinary.com/documentation/video_player_customization?utm_source=twilio&utm_medium=event&utm_campaign=cloudinary-twilioquest-2021#common_configuration_examples)
 
 ## Detailed Instructions
-In the links above, you will find code samples that will help you fill in the `<video>` tag of the video player embed. Using what you learned before about transformations, fill in the code and make your filled Flower video come to life!
-
-Your Flower video should still be located under `TwilioQuest/Flower`.
-
-For the title, subtitle, and description, put whatever you'd like that makes sense.
+Details
 
 ### Additional Documentation
